+++ packages/core/src/constants.ts
@@ -60,6 +60,7 @@ export const MODULES = [
   "futures",
   "option",
   "account",
+  "news",
   ...EARN_SUB_MODULE_IDS,
   ...BOT_SUB_MODULE_IDS,
 ] as const;